feat: move theme services to helpwave_theme package
