+++ packages/helpwave_theme/lib/theme.dart
@@ -1,2 +1,3 @@
-export "src/light_theme.dart" show lightTheme;
-export "src/dark_theme.dart" show darkTheme;
+export "package:helpwave_theme/src/light_theme.dart" show lightTheme;
+export "package:helpwave_theme/src/dark_theme.dart" show darkTheme;
+export "package:helpwave_theme/src/theme_model.dart";
