+++ packages/helpwave_theme/pubspec.yaml
@@ -10,6 +10,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  shared_preferences: ^2.0.15
 
 dev_dependencies:
   flutter_test:
